+++ src/Frontend/Dfe.Complete/Pages/Projects/AboutTheProject/AboutTheProject.cshtml.cs
@@ -1,9 +1,82 @@
+using Dfe.AcademiesApi.Client.Contracts;
+using Dfe.Complete.Application.Projects.Models;
+using Dfe.Complete.Application.Projects.Queries.GetProject;
+using Dfe.Complete.Application.Projects.Queries.GetTransferTasksData;
+using Dfe.Complete.Application.Services.AcademiesApi;
 using Dfe.Complete.Pages.Projects.ProjectView;
+using Dfe.Complete.Utils;
 using MediatR;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Dfe.Complete.Pages.Projects.AboutTheProject
 {
-    public class AboutTheProjectModel(ISender sender) : ProjectLayoutModel(sender, AboutTheProjectNavigation);
+    public class AboutTheProjectModel(ISender sender) : ProjectLayoutModel(sender, AboutTheProjectNavigation)
+    {
+        public EstablishmentDto? Academy { get; set; }
+        public ProjectGroupDto? ProjectGroup { get; set; }
+        public TransferTaskDataDto? TransferTaskData { get; set; }
 
-    //public class AboutTheProjectModel1(ISender sender) : BaseProjectTabPageModel(sender, AboutProject);
+        private async Task SetAcademyAsync()
+        {
+            if (Project.AcademyUrn != null)
+            {
+                var academyQuery = new GetEstablishmentByUrnRequest(Project.AcademyUrn.Value.ToString());
+                var academyResult = await sender.Send(academyQuery);
+
+                if (!academyResult.IsSuccess || academyResult.Value == null)
+                {
+                    throw new NotFoundException($"Academy {Project.AcademyUrn.Value} does not exist.");
+                }
+
+                Academy = academyResult.Value;
+            }
+        }
+
+        private async Task SetProjectGroupAsync()
+        {
+            if (Project.GroupId != null)
+            {
+                var projectGroupQuery = new GetProjectGroupByIdQuery(Project.GroupId);
+                var projectGroup = await sender.Send(projectGroupQuery);
+                if (projectGroup.IsSuccess || projectGroup.Value != null)
+                {
+                    ProjectGroup = projectGroup.Value;
+                }
+            }
+        }
+
+        private async Task SetTransferTaskDataAsync()
+        {
+            if (Project.TasksDataId != null)
+            {
+                var transferTasksDataQuery = new GetTransferTasksDataByIdQuery(Project.TasksDataId);
+                var transferTasksData = await sender.Send(transferTasksDataQuery);
+                if (transferTasksData.IsSuccess || transferTasksData.Value != null)
+                {
+                    TransferTaskData = transferTasksData.Value;
+                }
+            }
+        }
+
+        public override async Task<IActionResult> OnGet()
+        {
+            await SetProjectAsync();
+
+            await SetEstablishmentAsync();
+
+            await SetIncomingTrustAsync();
+
+            await SetOutgoingTrustAsync();
+
+            await SetAcademyAsync();
+
+            await SetProjectGroupAsync();
+
+            await SetTransferTaskDataAsync();
+
+            await SetCurrentUserTeamAsync();
+
+            return Page();
+        }
+    }
 }
