+++ src/Frontend/Dfe.Complete/Pages/Projects/AboutTheProject/_AboutTransferProject.cshtml
@@ -85,6 +85,8 @@
             @await Html.PartialAsync("Common/_AdvisoryBoardDetails", Model)
 
             @await Html.PartialAsync("Common/_AcademyDetails", Model)
+
+            @await Html.PartialAsync("Common/_IncomingTrustDetails", Model)
         </div>
     </div>
 </div>
